refactor: Move SetupCharacter to logics package and add dimension to Player
This commit refactors the `SetupCharacter` class by moving it from the `
utils` package to the `logics` package. This change improves code organization by grouping related classes together.

Additionally, a new `dimension` attribute has been added to the `Player` model. This attribute allows for specifying the dimensions of a character, which can be used for various purposes such as character positioning or
 scaling.